reanme XxxMock -> DraftXxx
