+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/lib/appservice/Draft.java
@@ -22,6 +22,6 @@
 
 package com.microsoft.azure.toolkit.lib.appservice;
 
-public interface Mock {
+public interface Draft {
     // marker interface
 }
